gui-wl: Use pointer constraints for warping
This is a neat trick borrowed from Ethan Burns' proposed plan9port wayland
backend that warps the pointer by locking the cursor, setting a position
hint, and then unlocking the cursor.

It works across compositors, and allows us to drop the dependency on
wlroots protocols.

Also fixed bugs in previous commit for handling close and events on the
decor surfaces.