gui-wl: avoid deadlock with clipboard
With how wayland works its possible that we're sending
updated clipboard content to ourselves. If so, we can't
have the main dispatch proc in charge of both; once we
fill the host pipe buffer we'll block on the write waiting
for a reader that will never show. Instead just spin out
a proc(pthread) when we're given someone to write to.

This happened to work because our max snarf size was below
the max of what the kernel would buffer for us. This was found
as part of bumping that limit.